requirements.txt
biopython==1.71
numpy==1.14.2
pandas==0.20.3
python-dateutil==2.7.2
pytz==2018.3
six==1.11.0
xmlschema==0.9.13
xmltodict==0.11.0
sh==1.12.14